Better error handling and comments
- Added better error handling for credit card details
- Added the ability to navigate back to previous sets incase the user miss clicks
- Added Comments describing what each method does, and how and why that a method does in areas that were not obvious

+++ BookStoreApp/src/bookstoreapp/Payment.java
@@ -20,7 +20,7 @@ public class Payment {
         
     }
     
-    public boolean takePayment(int cardNumber, YearMonth yearMonth, int cardCvv) {
+    public boolean takePayment(Long cardNumber, YearMonth yearMonth, int cardCvv) {
         // No payment serivce for the prototype. Using boolean to simulate taking payment 
         boolean paymentSuccessfull = true;
         return paymentSuccessfull;
